refactor: ファイル名をより一般的な名称に変更
