+++ app/lib/calculator.ts
